Emacs now asks the user about unsaved buffers before quitting
There is a slight trade-off here, in particular buffers not
associated with any file and even more particularly for *messages*
which is almost always modified without being saved but in general
the user won't wish to be asked about this. So the test for an
unmodified buffer includes a test for the buffer name in the case that
it begins and ends with a '*' and in that case we don't ask the user
about it.