Change modeApplies to look at contents of file.
The modeApplies function for Mode values now takes a parameter for the
contents of the file.  This allows for any information about the
contents of the file to be used for mode detection.  This record also
implements the use-case of looking into the contents for the shebang
directive of haskell script files. Do note that this can be buggy
because literate haskell scripts would also have the same shebang, so
the analyzer would have to be improved to distinguish the two styles.