Fix the GTK and Pango UIs to compile again.
Nicolas Pouillard's patch "Activate the reloading function under Mac OS X (still experimental)."
changed the type signature for UI.end, but did not update the GTK and Pango UIs to match
this new type. This is a trivial patch to fix that. The GTK and Pango UIs now
just ignore the new argument to UI.end.