Max# of entries in persistent state is configurable.
I added a configuration variable that limits the number of entries in two
types of in PersistentState: histories (per-command), and VimTagStack.

Size of an entry is not taken into account, since we don't expect
user to have excessively sized tags, nor commands.

[Killring size is limited to 10 during editor operation anyway,
and this is probably more likely to give an excessive memory
requirement, in case a user yanks a multimegabyte regions.
Still I believe that it is worth the added convenience,
when trying to move a fragment to the other file,
or restoring a recently deleted region.]