G without a count works correctly in operator pending mode
G needs to distinguish between having a count of 1 and having no count. This
requires the default value for no count to not be obtained when parsing the
events, but later when the motion interprets them.

The text objects and operators continue to take a default of 1 during the
command parsing.