Fix ~ behaviour at EOL
Introduce a transformCharactersInLineN function which takes an Int which is
the number of Chars to transform and a function Char -> Char which performs
the transformaton.

As this function uses a StyledRegion, which is defined in
Keymap.Vim2.StyledRegion, this can't be added to Buffer.Misc without
introducing a dependency for Buffer.Misc on Keymap.Vim2.StyledRegion.

Perhaps this is an indication that StyledRegion is sufficiently generic to be
moved somewhere under Buffer