Drop GHC <=7.4.x
Seeing as 7.8 is coming out and 7.6 has been out for a long time, we're
dropping 7.4. The motivation is that having to consider multiple
versions of ‘base’ and not being able to use no-longer-new language
extensions is not cool. You shouldn't really be hacking on current
Haskell software with a few-years-old compiler anyway.