Tests for Vim2 ex command parsers
These tests check two things:

 1. The parser successfully parses the commands it is expected to.
 2. For each sample string, the expected `ExCommand` is selected from the
    `defExCommandParsers`.

They don't check that any `ExCommand` has the affect that it should. Some of
the tests in `TestVim.getTestGroup` do that.

Generating the sample strings for each ex command parser is currently a manual
exercise. Which is why there are currently tests for only a few commands. It
should be possible to generate the command strings using SmallCheck or
QuickCheck.