QuickCheck generates the command strings to parse
Instead of manually generating a hopefully exhaustive set of command strings
to test, use QuickCheck to generate them.

Each ex command parser to test is now contained in a `CommandParser` record.
Along with the parser, the record contains all of the synonyms for the
command, whether the command accepts a `!`, and a `Gen`erator or command line
arguments it is expected to parse.

A number of combinators have been written to ease the generation of arguments.

Currently, only three parsers are tested. However, it should be simple to add
new ones.