Tests for pure-editor-manipulating Vim2 actions
The TestVim module provides a nice mechanism for writing tests which are 1)
pure and 2) affect only a single buffer. It doesn't provide a mechanism for
tests which affect more than a single buffer.

Pure manipulations of the editor refers to such things as changing layout,
navigating buffers, creating or deleting buffers, creating or deleting tabs.
In short, anything which 1) doesn't perform IO and 2) interacts with something
other than a single buffer.

This module provides the start of a basis for writing such tests.